+++ src/ops/queue.h
@@ -6,5 +6,21 @@
 extern const tele_op_t op_Q;
 extern const tele_op_t op_Q_AVG;
 extern const tele_op_t op_Q_N;
-
+extern const tele_op_t op_Q_CLR;
+extern const tele_op_t op_Q_GRW;
+extern const tele_op_t op_Q_SUM;
+extern const tele_op_t op_Q_MIN;
+extern const tele_op_t op_Q_MAX;
+extern const tele_op_t op_Q_RND;
+extern const tele_op_t op_Q_SRT;
+extern const tele_op_t op_Q_REV;
+extern const tele_op_t op_Q_SH;
+extern const tele_op_t op_Q_ADD;
+extern const tele_op_t op_Q_SUB;
+extern const tele_op_t op_Q_MUL;
+extern const tele_op_t op_Q_DIV;
+extern const tele_op_t op_Q_MOD;
+extern const tele_op_t op_Q_I;
+extern const tele_op_t op_Q_2P;
+extern const tele_op_t op_Q_P2;
 #endif