+++ CHANGELOG.md
@@ -25,6 +25,8 @@
 - **NEW**: v/oct to hz/v conversion op: `HZ`
 - **FIX**: W/2.0 ops added
 - **NEW**: W/2.0 ops documentation
+- **NEW**: `><`, `<>`, `>=<` and `<=>` OPs, checks if value is within or outside of range
+- **IMP**: new powerful Q OPs
 
 ## v3.2.0